Crate batch_system

Modules

batch
This is the core implementation of a batch system. Generally there will be two different kind of FSMs in TiKV's FSM system. One is normal FSM, which usually represents a peer, the other is control FSM, which usually represents something that controls how the former is created or metrics are collected.

config

fsm

mailbox

router

test_runner
A sample Handler for test and micro-benchmark purpose.

Structs

BasicMailbox
A basic mailbox.

BatchSystem
A system that can poll FSMs concurrently and in batch.

Config

Mailbox
A more high level mailbox.

Router
Router route messages to its target mailbox.

Traits

Fsm
A Fsm is a finite state machine. It should be able to be notified for updating internal state according to incoming messages.

HandlerBuilder
A builder trait that can build up poll handlers.

PollHandler
A handler that poll all FSM in ready.

Functions

create_system
Create a batch system with the given thread name prefix and pool size.

Type Definitions

BatchRouter