Module batch_system::batch

This is the core implementation of a batch system. Generally there will be two different kind of FSMs in TiKV's FSM system. One is normal FSM, which usually represents a peer, the other is control FSM, which usually represents something that controls how the former is created or metrics are collected.

Structs

Batch
A basic struct for a round of polling.

BatchSystem
A system that can poll FSMs concurrently and in batch.

ControlScheduler

NormalScheduler

Poller
Internal poller that fetches batch and call handler hooks for readiness.

Enums

FsmTypes
A unify type for FSMs so that they can be sent to channel easily.

ReschedulePolicy

Traits

HandlerBuilder
A builder trait that can build up poll handlers.

PollHandler
A handler that poll all FSM in ready.

Functions

create_system
Create a batch system with the given thread name prefix and pool size.

Type Definitions

BatchRouter