Trait batch_system::fsm::Fsm

pub trait Fsm {
    type Message: Send;
    fn is_stopped(&self) -> bool;

    fn set_mailbox(&mut self, _mailbox: Cow<'_, BasicMailbox<Self>>)
    where
        Self: Sized,
    { ... }

    fn take_mailbox(&mut self) -> Option<BasicMailbox<Self>>
    where
        Self: Sized,
    { ... }
}

A Fsm is a finite state machine. It should be able to be notified for updating internal state according to incoming messages.

Associated Types

type Message: Send

Loading content...
Required methods

fn is_stopped(&self) -> bool

Loading content...
Provided methods

fn set_mailbox(&mut self, _mailbox: Cow<'_, BasicMailbox<Self>>) where
Self: Sized,

Set a mailbox to Fsm, which should be used to send message to itself.

fn take_mailbox(&mut self) -> Option<BasicMailbox<Self>> where
Self: Sized,

Take the mailbox from Fsm. Implementation should ensure there will be no reference to mailbox after calling this method.

Loading content...