Struct batch_system::Router

pub struct Router<N: Fsm, C: Fsm, Ns, Cs> {
    normals: Arc<Mutex<HashMap<u64, BasicMailbox<N>>>>,
    caches: Cell<LruCache<u64, BasicMailbox<N>>>,
    pub(super) control_box: BasicMailbox<C>,
    pub(crate) normal_scheduler: Ns,
    control_scheduler: Cs,
    shutdown: Arc<AtomicBool>,
}

Router route messages to its target mailbox.

Every fsm has a mailbox, hence it's necessary to have an address book that can deliver messages to specified fsm, which is exact router.

In our abstract model, every batch system has two different kind of fsms. First is normal fsm, which does the common work like peers in a raftstore model or apply delegate in apply model. Second is control fsm, which does some work that requires a global view of resources or creates missing fsm for specified address. Normal fsm and control fsm can have different scheduler, but this is not required.

Fields
normals: Arc<Mutex<HashMap<u64, BasicMailbox<N>>>>caches: Cell<LruCache<u64, BasicMailbox<N>>>control_box: BasicMailbox<C>normal_scheduler: Nscontrol_scheduler: Csshutdown: Arc<AtomicBool>
Implementations

impl<N, C, Ns, Cs> Router<N, C, Ns, Cs> where
N: Fsm,
C: Fsm,
Ns: FsmScheduler<Fsm = N> + Clone,
Cs: FsmScheduler<Fsm = C> + Clone,

pub(super) fn new(
control_box: BasicMailbox<C>,
normal_scheduler: Ns,
control_scheduler: Cs
) -> Router<N, C, Ns, Cs>

pub fn is_shutdown(&self) -> bool

The Router has been already shutdown or not.

fn check_do<F, R>(&self, addr: u64, f: F) -> CheckDoResult<R> where
F: FnMut(&BasicMailbox<N>) -> Option<R>,

A helper function that tries to unify a common access pattern to mailbox.

Generally, when sending a message to a mailbox, cache should be check first, if not found, lock should be acquired.

Returns None means there is no mailbox inside the normal registry. Some(None) means there is expected mailbox inside the normal registry but it returns None after apply the given function. Some(Some) means the given function returns Some and cache is updated if it's invalid.

pub fn register(&self, addr: u64, mailbox: BasicMailbox<N>)

Register a mailbox with given address.

pub fn register_all(&self, mailboxes: Vec<(u64, BasicMailbox<N>)>)

pub fn mailbox(&self, addr: u64) -> Option<Mailbox<N, Ns>>

Get the mailbox of specified address.

pub fn control_mailbox(&self) -> Mailbox<C, Cs>

Get the mailbox of control fsm.

pub fn try_send(
&self,
addr: u64,
msg: N::Message
) -> Either<Result<(), TrySendError<N::Message>>, N::Message>

Try to send a message to specified address.

If Either::Left is returned, then the message is sent. Otherwise, it indicates mailbox is not found.

pub fn send(
&self,
addr: u64,
msg: N::Message
) -> Result<(), TrySendError<N::Message>>

Send the message to specified address.

pub fn force_send(
&self,
addr: u64,
msg: N::Message
) -> Result<(), SendError<N::Message>>

Force sending message to specified address despite the capacity limit of mailbox.

pub fn send_control(
&self,
msg: C::Message
) -> Result<(), TrySendError<C::Message>>

Force sending message to control fsm.

pub fn broadcast_normal(&self, msg_gen: impl FnMut() -> N::Message)

Try to notify all normal fsm a message.

pub fn broadcast_shutdown(&self)

Try to notify all fsm that the cluster is being shutdown.

pub fn close(&self, addr: u64)

Close the mailbox of address.